backend/requirements.txt
fastapi
sqlmodel
python-multipart
passlib
python-jose[cryptography]
uvicorn
bcrypt